refactor(wait_for): Changed throw_exc default to True
We've decided to change the default of throw_exc to True,
since when the function is used with throw_exc=False, and it reaches timeout,
there is next to no indication, other than a print message, that the wait_for has failed.
Also, this change aligns the default with the parallel function in dtest.