improvement(build-id on reports): adding Scylla build-id
there were many cases where the emails with report had
Scylla version, but without build-id. During perf jobs
while we bisect many issues and regressions, we had to
download either sct.log or db logs to find the build-id,
because if there were more than one build in the same day
the only way to identify the relocatable packages is using
the build-id.